Split library into submodules
