+++ textures.py
@@ -1,0 +1,8 @@
+#region TEXTURES
+''' NEXT UPDATE
+def create_texture(name, type):
+    pass
+def delete_texture(name):
+    pass
+'''
+#endregion
