+++ vertex_groups.py
@@ -1,0 +1,10 @@
+#region VERTEX GROUPS
+''' NEXT UPDATE
+def create_vertex_group(ref, group_name):
+    ref.vertex_groups.new(name=group_name)
+    return ref.vertex_groups[group_name]
+
+def delete_vertex_group(ref, group_name):
+    pass
+'''
+#endregion
